Rename MediaUploader component per our newer standard
